requirements.txt
openai
pandas
openpyxl
requests
mcp-playwright
python-dotenv
